Tests: fix compile warnings from GCC

+++ src/tests/repeat_walker_tests.c
@@ -78,7 +78,7 @@ static void test_repeat_loop()
   // Construct graph but no paths
   build_graph_from_str_mt(&graph, 0, seq, strlen(seq), false);
   TASSERT2(hash_table_nkmers(&graph.ht) == 15+12+15,
-           "%llu kmers", hash_table_nkmers(&graph.ht));
+           "%zu kmers", (size_t)hash_table_nkmers(&graph.ht));
 
   // Find first node in sequence
   dBNode node0 = db_graph_find_str(&graph, seq);
